jsrunner: Remove duplicated numberLines function definition
